* nasl/CMakeLists.txt: Add -Wextra flag.
* nasl/nasl_builtin_nmap.c (nse_script_destroy, simple_item_destroy)
(xml_start_element, xml_end_element, xml_read_text, xmltag_open_host)
(xmltag_open_ports, xmltag_open_cpe, xmltag_open_hostscript),
nasl/nasl_builtin_openvas_tcp_scanner.c (std_port, banner_grab)
(plugin_run_openvas_tcp_scanner), nasl/nasl_builtin_synscan.c
(packetdead, rm_dead_packets, extracttcp, v6_extracttcp, extractsport)
(issynack, mktcpv6, v6_sendpacket, scan, plugin_run_synscan),
nasl/nasl_crypto.c (nasl_get_sign, nasl_nt_owf_gen, nasl_lm_owf_gen),
nasl/nasl_grammar.y (naslerror), nasl/nasl_host.c (nasl_this_host_name),
nasl/nasl_http.c (cgibin), nasl/nasl_init.c (init_nasl_library),
nasl/nasl_isotime.c (nasl_isotime_now), nasl/nasl_misc_funcs.c
(nasl_rand, nasl_unixtime, nasl_dec2str, nasl_get_byte_order),
nasl/nasl_packet_forgery.c (dump_tcp_packet, get_udp_element)
(set_udp_element, dump_udp_packet, nasl_tcp_ping, nasl_send_packet),
nasl/nasl_packet_forgery_v6.c (dump_tcp_v6_packet, get_udp_v6_element)
(set_udp_v6_elements, dump_udp_v6_packet, nasl_tcp_v6_ping)
(nasl_send_v6packet), nasl/nasl_scanner_glue.c (script_summary)
(script_get_preference_file_location, nasl_scanner_status),
nasl/nasl_smb.c (nasl_smb_versioninfo), nasl/nasl_socket.c
(nasl_get_sock_info), nasl/nasl_ssh.c (next_session_id)
(nasl_ssh_connect, verify_session_id, nasl_ssh_close_hook)
(nasl_ssh_session_id_from_sock, exec_ssh_cmd_alarm)
(request_ssh_shell_alarm), nasl/nasl_tree.c (dump_cell_val, dump_tree)
(nasl_type_name), nasl/nasl_wmi.c (nasl_wmi_versioninfo): Fix
unused-parameter and sign-compare extra warnings.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@27515 423fd1db-d629-0410-8442-d21db03e70f4

+++ nasl/nasl_packet_forgery.c
@@ -748,8 +748,8 @@ dump_tcp_packet (lex_ctxt * lexic)
       int a = 0;
       struct ip *ip = (struct ip *) pkt;
       struct tcphdr *tcp = (struct tcphdr *) (pkt + ip->ip_hl * 4);
-      int j;
-      int limit;
+      unsigned int j;
+      unsigned int limit;
       char *c;
       limit = get_var_size_by_num (lexic, i - 1);
       printf ("------\n");
@@ -933,7 +933,7 @@ get_udp_element (lex_ctxt * lexic)
   char *udp;
   char *element;
   struct ip *ip;
-  int ipsz;
+  unsigned int ipsz;
   struct udphdr *udphdr;
   int ret;
 
@@ -999,7 +999,7 @@ tree_cell *
 set_udp_elements (lex_ctxt * lexic)
 {
   struct ip *ip = (struct ip *) get_str_local_var_by_name (lexic, "udp");
-  int sz = get_local_var_size_by_name (lexic, "udp");
+  unsigned int sz = get_local_var_size_by_name (lexic, "udp");
   char *data = get_str_local_var_by_name (lexic, "data");
   int data_len = get_local_var_size_by_name (lexic, "data");
 
@@ -1119,9 +1119,9 @@ dump_udp_packet (lex_ctxt * lexic)
   while ((pkt = (u_char *) get_str_var_by_num (lexic, i++)) != NULL)
     {
       struct udphdr *udp = (struct udphdr *) (pkt + sizeof (struct ip));
-      int j;
+      unsigned int j;
       char *c;
-      int limit = get_var_size_by_num (lexic, i - 1);
+      unsigned int limit = get_var_size_by_num (lexic, i - 1);
       printf ("------\n");
       printf ("\tuh_sport : %d\n", ntohs (udp->uh_sport));
       printf ("\tuh_dport : %d\n", ntohs (udp->uh_dport));
@@ -1377,7 +1377,7 @@ nasl_tcp_ping (lex_ctxt * lexic)
   struct in_addr src;
   struct sockaddr_in soca;
   int flag = 0;
-  int i = 0;
+  unsigned int i = 0;
   int bpf;
   char filter[255];
   tree_cell *retc;
@@ -1544,7 +1544,7 @@ nasl_send_packet (lex_ctxt * lexic)
       int sz = get_var_size_by_num (lexic, vi);
       vi++;
 
-      if (sz < sizeof (struct ip))
+      if ((unsigned int) sz < sizeof (struct ip))
         {
           nasl_perror (lexic, "send_packet(): packet is too short!\n");
           continue;
